services/report-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
pandas==2.1.0
openpyxl==3.1.2
reportlab==3.7.0
sentry-sdk==1.11.0
opentelemetry-instrumentation-fastapi==0.51.0
starlette-exporter==0.23.0
